lib: Add a GsCategoryManager
This is a single place which keeps track of the `GsCategory` objects and
allows looking them up, which ensures that if one plugin modifies a
`GsCategory` object (for example, to change the number of apps in it),
all other plugins will see that change.

It allows decoupling the enumeration of categories (which should always
be sync, since they’re coming from a static array) from refining them
(which should be async, as it depends on data loaded by plugins).

This commit adds the category manager, but doesn’t use it yet.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>