gs-category: Handle size of ‘all’ subcategory specially
Rather than setting the size of all the ‘all’ subcategories in a second
pass over the categories when creating them, handle the ‘all’
subcategory specially in `gs_category_get_size()`, so that its return
value is always correct.

This will allow for elimination of the second pass in subsequent
commits.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>