gs-plugin-loader: Make final rather than derivable
This simplifies the code a little by reducing boilerplate. It seems
unlikely the plugin loader is going to be subclassed at any point.

The change from `priv->` to `plugin_loader->` was done with find and
replace, then all calls to `gs_plugin_loader_get_instance_private()`
were dropped, and definitions of `plugin_loader` readded in a couple of
places where it wasn’t already defined.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>