gs-updates-page: Fix margins and spacing across page
Previously there was no margin on the page itself, so when rendered in a
narrow container, the borders of the updates sections would have no
margin.

The top margin of the page was previously provided by the top margin of
the topmost updates section. The bottom margin was provided separately
and didn’t match. Rework that so that the page’s margin is set at the
page level, and the spacing is set to separate the sections from each
other.

This makes the updates page behave more like the installed apps page for
margins.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>