src: Move time-to-string function to gs-common
This will allow it to be reused in later commits. Also change it to be
more in line with how time spans are displayed in the mock-ups for the
version history widget, e.g. "5 weeks ago" instead of "25 May 2012".
This code is based on code in gnome-shell's js/misc/util.js.