gs-featured-carousel: Add new carousel widget for featured apps
This is a new design for showing featured apps, using `HdyCarousel`.
Eventually, it will prioritise an app-specific background colour, the
app’s hi-res icon, and some text; instead of the current approach of
supporting custom per-app CSS. The CSS was hard to implement, the
background images took up a lot of memory and disk space, and it
required a lot of curation time which it turned out we just didn’t have.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1131