plugins: Move key-colors plugin into GsApp code
There’s no need for alternate implementations of the key-colors plugin,
or for it to be disabled conditionally, so it’s more performant to move
the plugin into the main library. That reduces the number of plugins to
iterate over for each vfunc call.

The key-colors code itself is unchanged, just plumbed in to
`gs_app_get_key_colors()` to be calculated on demand. The `key_colors`
array is now created lazily.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>