plugins: Move key-colors-metadata plugin into GsApp code
There’s no need for alternate implementations of the key-colors-metadata
plugin, or for it to be disabled conditionally, so it’s more performant
to move the plugin into the main library. That reduces the number of
plugins to iterate over for each vfunc call.

The key-colors-metadata code itself is unchanged, just moved into a
helper method in `gs-common.c`, which code using CSS metadata from apps
can call when needed.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>