gs-app: Rename gs_app_set_pixbuf() to gs_app_add_pixbuf()
This is a step towards supporting multiple icon sizes for apps, so that
different parts of the UI can display icons at different sizes.

Backends need to be able to expose all the natively-supported icon sizes
for an app, rather than being forced to provide the largest-possible
icon and have it downsized for most use cases.

Subsequent commits will rework the rest of the `GsApp` pixbuf/icon API.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>