gs-app: Provide pixbufs at any requested size
Rework the `gs_app_get_pixbuf()` API to return a pixbuf at a requested
size, supporting multiple stored pixbufs per app, and resizing them in
the `GsApp` code when needed, rather than making the caller do it every
time.

This allows more than one pixbuf for a `GsApp`, which allows the updated
`GsFeatureTile` to use a different pixbuf size (128px) from the rest of
the interface (64px).

In future, it should be possible to fold the `gs-plugin-icons` code into
`GsApp`, so that the metadata in `AsIcon` about *where* a pixbuf came
from isn’t lost, which should allow pixbufs to be dropped out of memory
and transparently reloaded when next needed, lowering runtime memory
usage. However, that is future work.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1131