gs-debug: Split out environment variables from GsDebug state
Allow the `GsDebug` state to be initialised from the environment
variables, but also allow other initialisation sources. This should make
unit testing a little easier, and reduce reliance on `setenv()` (which
is not thread-safe).

The internal mutex can be dropped and atomic variables used for anything
which might change at runtime, instead. (See following commits which
make use of this.)

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>