gs-debug: Allow setting verbose mode at runtime
Add a new function to set verbose mode, and hook it up in `gs-cmd` and
`gs-application` to the `--verbose` command line arguments.

In future, it should be possible to hook it up to a D-Bus interface
which would allow verbose mode to be enabled dynamically at runtime.
That might make debugging a bit easier.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>