lib: Use GsDebug in tests to avoid calling setenv() at runtime
`setenv()` is not thread-safe, and its usage here could destroy the
`environ` array for all variables if we’re unlucky. That could cause
spurious test failures.

This includes a change to a test string match, as the `GsDebug` output
format is slightly different from the default GLib one (which is what
the test used previously).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>