Adjust activatability of add-on rows
Currently if you hover your cursor over a GsAppAddonRow, it dims, since a
GtkListBoxRow is activatable by default. But clicking it does nothing, so
dimming is the wrong behavior. Make it so that clicking an add-on row that is
not installed is equivalent to checking the checkbox to install it (and
clicking an add-on row in the "queued for install" state un-checks the
checkbox). When the add-on is installed and has a "Remove" button, make the row
unactivatable; the user should have to explicitly click the button in that
case.

Relatedly, make the row unselectable and make the list box unfocusable, which
are in line with how the version history list box works.

https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/633