gs-updates-page: Keep showing apps which are being installed
When a user enters the Updates page and sees one Flatpak application for update, then
he/she clicks "Update" and the application switches to the "installing" state and the user
can watch the progress of the install. While the page is shown it can either refresh
on its own or after user's intent to refresh the content. When it's done the page shows
empty with "System is up to date", which is considered a lie, because there is still
an ongoing update. This can be verified in the Installed page, where the application
is shown as "installing".

Rather than presenting the user suddenly empty Updates page, show there also applications,
which are installing.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1150