lib: Rework key color extraction code
Rework it so that it uses a standard k-means algorithm, and does so
without doing any allocations. Using the `profile-key-colors` test
program, this reduces runtime for extracting key colors from an average
icon from
[411, 7189]μs, mean 2379±986μs, n = 1038
to
[200, 774]μs, mean 304±60μs, n = 1038

Note that the new code extracts up to 3 key colors, whereas the old code
extracted up to 10. If the old code was limited to 3 key colors, it was
about half the speed of the new code, rather than an eighth. Only
extracting 3 colours is justified by looking at a selection of icons:
the modal number of key colours (subjectively) needed is 3.

See the new comments in `gs-key-colors.c` for more details about the new
algorithm and the decisions behind it.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1148