gs-feature-tile: Choose the best key colour as the background colour
Add a new algorithm to sort the app’s key colours by their suitability
for use as background colours, and use the most suitable one. If none
are particularly suitable, tweak one of them so it is.

See the comments in `gs-feature-tile.c` for details of the algorithm and
justifications for why it was designed that way.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1148