gs-feature-tile: Recalculate background colour when style changes
This will happen if the user changes the desktop theme while a feature
tile is visible. That’s expected to be rare, but possible. More
typically, this will happen when a developer is testing gnome-software
with `GTK_DEBUG=interactive` and they toggle dark mode on/off. Useful
for testing.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1148