gs-app: Split out functionality to remove all icons
If it’s controlled by passing `NULL` as the new icon to
`gs_app_add_icon()`, then it could easily be triggered if a plugin fails
to load the icon it subsequently passes in. That seems like quite a big
consequence for a failure (even if it is a programmer error due to not
checking for errors properly).

Split out the functionality to remove all icons, so it’s explicitly a
separate function.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1147