gs-utils: Make creating cache directory optional
Add a new flag to `GsUtilsCacheFlags` which specifies to create the
cache directory, and change the default to *not* create the directory.
This means that, with appropriate flags, `gs_utils_get_cache_filename()`
can guarantee to not set an error, which means it can be called from
functions which cannot fail (or which cannot propagate an error).

Change all the existing call sites to pass the new flag
(`GS_UTILS_CACHE_FLAG_CREATE_DIRECTORY`) where it makes sense to do so.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1147