lib: Add GsRemoteIcon as a GIcon implementation for remote icons
This derives from `GFileIcon` (in an ugly way; see the comments in the
code) to provide a `GIcon` implementation and caching support for icons
which are at HTTP or HTTPS URIs.

It uses a predictable caching location for the icon so that it can be
referenced as a `GIcon` independently from when it’s actually downloaded
and cached.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1147