lib: Port GsApp to use GIcons rather than GdkPixbufs for icons
This moves the process of loading an icon (as a `GdkPixbuf`) from the
time when the `GsApp` is created, to the time when the icon is displayed
in the UI. This allows the pixbufs to be kept in memory only as long as
they’re displayed and, in particular, to not consume memory while the
main window of gnome-software is closed.

GTK itself maintains an LRU cache of icon pixbufs, so if a particular
icon is needed frequently, it shouldn’t end up being reloaded from disk
unnecessarily.

This commit uses various classes which implement `GIcon` to wrap
different icon sources which were previously all represented by
`AsIcon`. It wouldn’t have made sense to implement `GIcon` in `AsIcon`
itself — as well as being a minor layering violation, this approach
would not have supported other icon sources. In particular, flatpak
bundles provide their icons as binary blobs, rather than `AsIcon`s;
these need to be supported too. With this approach, they can be, using
`GBytesIcon`.

Another advantage of using `GIcon`s is that they allow multiple sizes of
icon to be loaded, rather than relying on loading the
most-commonly-requested size to a pixbuf, or loading the largest size
and scaling it down on command.

Remote icons (with a HTTP or HTTPS URI) are implemented using
`GsRemoteIcon`, introduced in a previous commit.

This allows for most of the implementation of `gs-plugin-icons` to be
dropped, as converting a `GIcon` to a `GdkPixbuf` is now mostly handled
by `gtk_image_set_from_gicon()`. The downloading and caching of remote
icons is handled in `GsRemoteIcon`, and `gs-plugin-icons` now remains
only to schedule calls to `gs_remote_icon_ensure_cached()` from a worker
thread. Eventually this too can disappear, but it’s a convenient way of
getting the threading right for now.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1147