lib: Drop gs_app_load_pixbuf() and use GIcons instead
This simplifies the code a bit, but significantly allows for
`GtkImage` to automatically reload icons if the icon theme changes, as
it now has access to icon names (from the `GIcon`) rather than just
already-loaded pixel data (in a `GdkPixbuf`). It also means the RTL
variant of an icon can be loaded if in an RTL locale (and if the variant
exists; I can find no apps which currently take advantage of this
functionality).

This doesn’t matter for most application icons, but could do if an icon
theme distributed overridden versions of popular app icons.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1147