gs-flatpak: Export both sizes of icons from flatpak bundles
Flatpak currently supports exactly 64px and 128px icons in bundles, and
no other sizes. Add both of them to the `GsApp` (if available), so that
the icon choosing code in `GsApp` has more source material to choose
from.

This means that `GsApp` will more likely be able to choose an icon at a
requested size without having to scale any pixbufs.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1147