gs-plugin-loader: Reset online-updates-timestamp only when did the update
The gs_plugin_loader_generic_update() is used in two cases, when updating
the software and when downloading the software. The former case can reset
the online-updates-timestamp, but the later not, because it did not
update anything.

This blocks the update notifications from showing up after the first
run of the Software.