gs-app: Use the libappstream function to check proprietary licenses
This should have equivalent behaviour, but saves us a bit of code and
maintenance.

Notably, it handles `LicenseRef-proprietary` with the same case as our
existing code, so this change doesn’t fix the underlying problem with
issue #1161. However, the underlying fix is for Steam Link to change its
appdata, and for the appstream spec to be consistent in its casing.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1161