gs-appstream: Fix handling of icons where appdata doesn’t specify width
If the `<icon>` element in the appdata doesn’t specify a `width`
attribute, `xb_node_get_attr_as_uint()` will return `G_MAXUINT64`.
However, its return value was being assigned to `sz`, which is a
`guint`, and hence it was being truncated. This resulted in icons having
their widths set to `G_MAXUINT`, and hence being prioritised for display
when that was not necessarily correct.

Fix that by explicitly handling the failure response from
`xb_node_get_attr_as_uint()`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1171