gs-icon: Work around invalid icon names from some appstream files
Cached icons are supposed to provide a name (like `krita.png`), `width`
and `height` attributes, and then gnome-software can build the path to
the cached icon file according to
https://www.freedesktop.org/software/appstream/docs/sect-AppStream-IconCache.html#spec-iconcache-location.

Due to a bug in appstream-builder in appstream-glib
(https://github.com/hughsie/appstream-glib/pull/390), some appstream
files – for example, rpmfusion-free-33 – had invalid cached icon names,
which resulted in those icons not being loadable, and not appearing for
their applications.

As it may take a while for the fix in appstream-builder to be merged and
be used to regenerate all the appstream files for different app
repositories, add a workaround in gnome-software.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1171