gs-installed-page: Hide the removable apps header if shown first
This restores the behaviour for this header from before the previous
commit, but allows this header to be shown if it’s *not* the first
section in the list.

In particular, this can happen if some apps are currently being
installed or removed, as they are now sorted to the top of the list,
in the ‘In Progress’ section.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1175