flatpak: Allow cancelling querying update permissions
Sometimes, calling `flatpak_installation_fetch_remote_metadata_sync()`
can result in HTTP downloads which are not small. It’s called as part of
searching for updates, and so a cancel button is available for it on the
‘Updates’ page in the UI.

If the user tries to cancel searching for updates while the flatpak
plugin is in a call to `gs_flatpak_set_update_permissions()`, nothing
will happen and the UI will continue to show that updates are being
searched for.

Fix that by passing a `GCancellable` in, so that the permissions updates
can be cancelled promptly.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>