flatpak: Refresh UI after changes made outside of Software
When the flatpak installation monitor claims it changed in the background,
the memory plugin cache of the GsApp should be discarded and the information
about installed/to-be-updated applications refreshed as well. The UI should
be reloaded, to reflect the actual state, instead of showing stale data, aka
the GsApp, which had been in the cache.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/630