Speed up search queries by over 40%
Use the built-in tokenization support in libxmlb so we can mmap the search
tokens from the silo rather than splitting them for each query.

This also reduces the amount of peak heap memory used, and the number of heap
allocations by several orders of magnitude for each query performed.

This requires building against a very new version of libxmlb to work.