gs-details-page: Deduplicate the origin list
If an app has been renamed, or provides the same `.desktop` file as
another app, and has the same origin as the original app, that origin
will appear twice in the origin list.

Avoid that by deduplicating the list on the fields which are presented
in UI.

This situation won’t happen often (it’s an unlikely set of metadata to
have, although possible). It’s reproducible with
Chromium/UngoogledChromium though.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1192