GsAppRow: Emit "unrevealed" signal only if mapped
There can happen, with proper timing, that a GsAppRow is being revealed
from the Installed or Updates page when the page reloads and removes
all rows from itself. The "unrevealed" signal handler also removes
the row from the list and it is emitted when the list unmaps the row,
as a consequence of the end of the reveal animation, it can cause a crash
due to NULL parent property dereference on the gtk+ side.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1195