rpm-ostree: Reference per-transaction objects and use them for the whole transaction
Rather than re-using semi-global D-Bus proxies, reference them at the beginning
of the transaction call and use the same for the whole transaction. This is
a preparation for eventual free of the objects when being inactive for a long time,
which can happen during a long standing transaction (when the transaction runs
longer than the inactive timeout is set).