gs-shell: Overview page opens empty with "Automatic Updates" off
When the "Automatic Updates" are off, the shell window switches to
the overview page immediately. Since some recent change the preselected
page is the Overview page, thus the switch to it doesn't call reload(),
thus it is left empty. To adapt to the new behavior call reload() also
when the Overview page had been already selected. It's okay, because
it's only after the initial start.

Non-trivial cherry-pick of 80fc7459 to `gnome-40`.

Fixes: #1070