gs-search-page: Reload the page on the text change when the page is active
This fixes the initial search when the page is selected before the text
is set. The text set only marked the need to reload itself, but it relied
on the page change signal to do it. As the page is already active, the signal
did not come and no search was done.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1185