ci: Add an on-demand flatpak bundle build
As mentioned in `README.md`, this is not an official release, nor is it
fully functional (only the flatpak plugin is built). It will be useful
for testing upcoming UI changes without requiring designers to build
gnome-software themselves.

You can build it by manually triggering the ‘flatpak bundle’ job in any
CI pipeline. Download `gnome-software-dev.flatpak` from the job’s
artifacts once it completes, and install and run using:
```
$ flatpak install --bundle ./gnome-software-dev.flatpak
$ flatpak run org.gnome.SoftwareDevel
```

The CI machinery uses
https://gitlab.gnome.org/GNOME/citemplates/-/blob/master/flatpak/flatpak_ci_initiative.yml
as a template, and is based on Nautilus’ use of the same
(https://gitlab.gnome.org/GNOME/nautilus/-/blob/master/.gitlab-ci.yml,
https://gitlab.gnome.org/GNOME/nautilus/-/blob/master/build-aux/flatpak/org.gnome.NautilusMaster.yml).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1206