Export desktop file & icon in development builds
Previously, no desktop file, app icon or D-Bus service would be
exported, because org.gnome.Software is not within the
org.gnome.SoftwareDevel namespace of the Flatpak.

With this change, the development build shows up in the app grid &
search and on the dash, with the "devel" icon already present in the
source tree.