shell: Split the modes in decks
This split the main view and the sub views into two stacks, and uses a
deck to move between the main and sub views with an "over" transition
animation.

In practice, this avoids weird jumps caused by changing the sidebar's
visibility, and this gives a better sense of spatialization to the user
that is reflected into the UI's layout.

In the future, it will allow to simplify the shell's code further and to
support gestures.

The indentation is purposefully left broken here to make the actual code
change easier to review and will be fixed in the next commit.

Fixes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1219