gs-category-page: Decouple from GsShell
Add a signal for when an app is clicked, and connect to that in the
`GsShell` to show an app’s details — rather than keeping a pointer to
the shell in `GsCategoryPage` and calling a method on it directly.

This removes a pointer loop between `GsShell` and `GsCategoryPage`, and
makes the category page more reusable in future.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>