data: Add a custom featured colour for Steam
Steam appears to be one of the corner cases where the built-in
heuristics for extracting key colours don’t manage to pick ones which
have enough contrast with the foreground text colour.

Rather than fiddle with the heuristics and potentially break them for
other apps (i.e. requiring a lot of testing), add an override key colour
for Steam.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1223