gs-category-page: Refactor queries to simplify code
This refactors the queries so that they are both done in parallel, and
the UI is only updated once when both queries have completed. The
featured apps query is now used to filter the main query results, rather
than being added to the UI independently.

This should all help to eliminate the possibility of races when updating
the UI, and will make it easier to list ‘recently updated’ apps in the
next few commits.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1111