gs-details-page: Explicitly handle empty version history array
This makes the code a bit more robust.

It’s not actually needed, as the current behaviour of
gs_app_get_version_history() is to either return `NULL` or a non-empty
array — never an empty array. However, guaranteeing that behaviour would
defy precedent in other APIs, and be confusing to document, so just
handle the case of empty arrays in the callers anyway, to make their
expected behaviour clear.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>