packagekit: Set interactive status on PkClient wherever it’s used
In the same way the background state and cache age are set. Note that
the background state was previously perhaps a little confused with the
interactive state. The background state indicates that an operation
should be allocated few system resources, as it’s not urgent. The
interactive state indicates that the user explicitly requested an
operation, and hence can be presented with polkit prompts (for example)
when necessary.

Ideally the packagekit-glib2 API would take this as a flag on each
operation, but instead it has to be set on the `PkClient`. This is
translated into a transaction hint whenever the `PkClient` creates a new
D-Bus proxy (which happens at the start of each operation).

Ideally also the internal gnome-software API would provide this as a
per-operation flag, but instead it’s set as per-plugin data (and
sometimes updated). Perhaps we can improve that in the future.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #582