gs-details-page: Enhance the alternates deduplication cycle
The deduplication works as long as the array doesn't contain the same pointers
or when those are not interleaved. Being interleaved it can happen the first
occurrence of the second or later instance is skipped due to the array content
shift with g_ptr_array_remove(), which removed the first occurrence of the data.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1192