gs-plugin-repos: Special-case when rpm repo name doesn't match appstream origin
For example rpmfusion uses "rpmfusion-free" as the repo name, but the appstream's
origin is set to "rpmfusion-free-NN", where the NN is a Fedora version number.
Try to catch such cases, thus the origin is known and the details page shows
the origin URL (without it the URL is not shown).