ci: Make the deploy CI job manual as well as the flatpak bundle job
Otherwise the CI pipeline stays in the ‘running’ state forever, as the
prerequisite for the deploy job has not been run (because it’s manual).
Note that this only applies to pipelines run on `master` as the deploy
job doesn’t run on merge requests.

I can’t think of a better way to fix this, but there probably is one.

Follows on from: https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/708

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>