flatpak: Update OSTree branch name
OSTree has renamed its development branch from 'master' to 'main'. They
have kept the 'master' branch around for the benefit of tools that
require pinned commits to be on the 'master' branch, but removed all
content from that branch.

As a result of this change, this Flatpak build fails, at least on my
system which had an existing checkout of the 'master' branch. I suspect
that on a new checkout, flatpak-builder would respect the server's
default branch, but being explicit here does not cause harm.

Explicitly specify the 'branch' as 'main' for the ostree module.

https://github.com/ostreedev/ostree/pull/2358
https://github.com/ostreedev/ostree/issues/2360