gs-odrs-provider: Split out from ODRS plugin
This keeps the ODRS code the same, making the minimal changes necessary
to package it up in a `GObject` and expose the `gs_plugin_*()` vfuncs as
methods on that object.

The new `GsOdrsProvider` class is then used in the ODRS plugin so that
behaviour is unchanged.

This is one step towards refactoring the code so the ODRS plugin can be
removed, by moving ODRS functionality into the core of gnome-software
itself.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>