gs-details-page: Use GsOdrsProvider directly
Rather than making ODRS requests via the plugin loader (which, for
example, calls the `review_upvote()` vfunc on all plugins in order to
hit the (only) implementation in the ODRS plugin), call into the
`GsOdrsProvider` directly.

This will make each ODRS call a little faster, which is not very
important. What is more useful is that it will allow a lot of vfuncs to
be removed from the plugin loader, which is a helpful simplification.

One behavioural change here is that the ODRS requests are now made
synchronously rather than asynchronously. This will change again with
future refactoring.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>