external-appstream: Move plugin into GsPluginLoader
The `external-appstream` plugin is the only one which is ordered before
the `appstream` plugin in the plugins graph. By removing it, it becomes
possible to move the `appstream` plugin itself into the core of
gnome-software. In turn, this will reduce interactions between plugins,
which will make the plugin threading model easier to modify in future.

The `external_appstream` Meson option continues to allow the external
appstream functionality (and `gs-install-appstream` helper binary) to be
disabled as before.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>