gs-plugin: Bind initialize/destroy to the GsPlugin lifetime
This is a step towards having a `GsPlugin` subclass per plugin, which
will allow the vfuncs to be class methods, and avoid a load of `dlsym()`
calls on every plugin job. It will also allow the C type system to be
used, rather than a load of dynamic function pointer casting in
`gs-plugin-loader.c`. Finally, it will prevent the perennial issue of
identically-named symbols from multiple modules being conflicting in the
process’ function name space, and being difficult to grep in the source
code.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>