gs-app: Add G_PARAM_STATIC_STRINGS to all properties
This prevents `GObject` from taking copies of the property name and
description.

Although almost all the properties don’t specify a name or description,
it’s good to specify this anyway, in case strings are added in future,
or the code is copy/pasted somewhere and strings are added there.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>