updates-page: Improve the "updates loading" page style
This better implement the HdyStatusPage style for the updates loading
state, splitting the page's title and its description in two strings,
using the title case for the title and making the description a proper
sentence.

This also drops gs_updates_page_get_state_string() as it's only used to
get a localized static string, making the code simpler and more
consistent.