progress-button: Allow showing an icon
This adds the icon-name and show-icon properties (and matching
accessors) and use them to allow showing an icon instead of a label,
with a smooth transition between the two thanks to a stack.

This also overrides the label property and add new accessors, as the
ones from GtkLabel destroy the widget's layout and hence can't be used.