gs-safety-context-dialog: Add safety context dialogue
This presents information about the safety/trustworthiness of an app to
the user.

A future commit will make it appear when the safety tile in
`GsAppContextBar` is clicked.

Some of the code for working out the safety is copied from the
`GsAppContextBar`. It will be refactored to remove the duplication in a
future commit.

Includes significant work by Adrien Plazas.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1111