gs-hardware-support-context-dialog: Add hardware support context dialogue
This presents information about what hardware the app supports/requires,
to the user.

A future commit will make it appear when the hardware tile in
`GsAppContextBar` is clicked.

Some of the code for working out the hardware support is copied from the
`GsAppContextBar`. It will be refactored to remove the duplication in a
future commit.

Includes significant work by Adrien Plazas.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1111