gs-age-rating-context-dialog: Add age rating context dialogue
This presents information about what ages the app is suitable for, to
the user.

A future commit will make it appear when the age rating tile in
`GsAppContextBar` is clicked.

Some of the code for working out the age rating is copied from the
`GsAppContextBar`. It will be refactored to remove the duplication in a
future commit.

Includes significant work by Adrien Plazas.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1111