gs-details-page: Refresh the 'Launch' button on app info changes
It can happen that after an application installation the GDesktopAppInfo
cache is not updated before the details page is refreshed, which causes
the 'Launch' button not being shown, even the application can be launched.
Let the button visibility be updated on any app info changes, which
is the time when also the GDesktopAppInfo cache is updated.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1265