feature-tile: Use body class for subtitle
The caption style is smaller than the style used for normal app tiles,
and looks way too small next to the large app name.

With this change, apps in the carousel have a larger name than in a
regular tile, and the description is the same size in both places.