gs-app: Reduce allocations for storing URLs
Lazily allocate the URLs hash table (which avoids an allocation for a
`GsApp` if the URLs are never refined), and index the hash table by
`AsUrlKind` directly rather than a copy of its string form. This should
save a lot of string duplications.

From a very quick test, this appears to reduce memory usage by 2MB.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>