license-tile: Tweak description for proprietary tile
In my opinion, sandboxing is our main defence against insecure or
nefarious apps (regardless of license), and Software already has a
separate indication of whether an app is well-sandboxed.

With that in mind, adjust the second sentence to reference the absence
of freedom zero ("The freedom to run the program as you wish, for any
purpose") and otherwise emphasise the inability to audit the app.

https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1349