gs-details-page: Rearrange the header to match the new designs
Designs: https://gitlab.gnome.org/Teams/Design/software-mockups/-/raw/master/adaptive/app-page-adaptive.png

This doesn’t implement changes to the buttons themselves, which are
going to be handled in #1351.

It moves the ratings below the app author, adjusts spacing, and moves
the buttons into a new `GtkBox` with the title/author name. In future,
this box should be easily replaced with an adaptible widget which could
re-orient to displaying its children vertically when narrow. That should
correctly handle the narrow version of the design. (Correctly and fully
implementing that is out of scope for GNOME 41 though.)

The indentation in the UI file will be fixed in a subsequent commit, to
make this one easier to review.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1350