gs-details-page: Origin/source picker sometimes hidden
There could be sometimes hid the origin/Source picker in error, when
the set app changes its state, which called, on idle, the switch_to(),
where the picker is hid. The switch_to() doesn't make sense there,
because the page is already selected, thus rather only refresh
the page content.