gs-plugin-appstream: Convert component kind to appstream name
The appstream-glib uses different name for web-application, which
the appstream doesn't know, thus rather convert the kind in the opposite
way, since the appstream is used these days. This convert is done
just in case there are old appstream data files.

Also drop the 'console' conversion, which looks like a leftover.