gs-details-page: Ensure large context tile lozenges are square
If a context tile lozenge contains an icon which is taller than it is
wide, and has the `large` class applied, its width is only constrained
by the icon and so it will end up narrower than it is tall.

Fix that by setting a min-width for large lozenges such that they’re
always at least 60x60px.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>