gs-shell: Fix keyboard shortcut for back button
Since rearranging the header bar, there are actually two back buttons,
shown in different header bars at different times.

The key press handling code was only activating one of them, so pressing
Alt+Left (for example) worked when on a category page, but not when on a
details page.

Activate both back buttons so that it works on all pages. GTK will
ensure that only the back button which is visible is actually activated.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>