gs-appstream: Treat all territorial variants of English the same
Assume that all territorial variants of English are mutually intelligble
to the extent that apps are unlikely to be translated to them, and it’s
not worth pointing out to the user that an app will only be in US
English and not British English (or Zimbabwean English, etc.).

If apps start being translated to English variants more commonly, we
could revert this commit.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1406