gs-upgrade-banner: Only show border on bottom half of banner
Allow the background image from the distro to provide the border on the
upper half.

The main change in this commit is to split the banner in half
vertically, so that separate CSS can be applied to the top and bottom
halves.

The default CSS (if the distro doesn’t override it) for the top half
continues to set the border — but the override CSS provided by the
distro now unsets it.

This commit can be tested using `GS_SELF_TEST_DUMMY_ENABLE=1
gnome-software`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1407