gs-appstream: Handle self-closing elements in descriptions better
Some oddly formatted appdata can contain descriptions like this one from
‘google Roboto Slab’:
```
<description><p/><p/></description>
```

After some testing, it appears that Firefox treats `<p/>` as if it
doesn’t exist, and `<li/>` the same as `<li></li>` — so that’s what we
should do.

These changes mean that gnome-software no longer displays
`(null)\n\n(null)` as the description in the UI for google Roboto Slab.
The description is now hidden for that font.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>